requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.2
pydantic[email]==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
google-genai==1.52.0